Precaution extension break on unset level
If the SARIF output doesn't have a value for level on
the result object, the extension fails.

This happens because SARIF will not output defaults for values
in some cases. If the result level is unset, then the value should
be retrieved from the rule default configuration level. If that
is also unset, the value should be warning.

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>